@@ -24,27 +24,28 @@
     "@ckb-lumos/codec": "next",
     "@ckb-lumos/lumos": "next",
     "@ckb-lumos/molecule": "next",
-    "@docusaurus/core": "^2.4.1",
-    "@docusaurus/preset-classic": "^2.4.1",
-    "@docusaurus/theme-common": "^2.4.1",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
+    "@docusaurus/theme-common": "^3.0.0",
     "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11.9.3",
-    "@mdx-js/react": "^1.6.22",
+    "@mdx-js/react": "^3.0.0",
     "@mui/icons-material": "^5.10.3",
     "@mui/material": "^5.9.1",
     "clsx": "^1.2.1",
     "numeral": "^2.0.6",
     "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-json-tree": "^0.17.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.1",
-    "@tsconfig/docusaurus": "^1.0.5",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@tsconfig/docusaurus": "^2.0.2",
     "@types/numeral": "^2.0.2",
-    "eslint-plugin-react": "^7.30.1",
-    "typescript": "^4.7.4"
+    "@types/react": "^18.2.38",
+    "@types/react-dom": "^18.2.17",
+    "eslint-plugin-react": "^7.30.1"
   },
   "browserslist": {
     "production": [
@@ -59,6 +60,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
