@@ -7079,8 +7079,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x92b197aa1fba0f63633922c61c92375c9c074a93e85963554f5499fe1450d0e5',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:  {
@@ -7115,8 +7117,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x7bb56e1288a1de98bab23d3e0ec7728634b6626ab03cc119ec23005a82ff12ff',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:  {
@@ -7151,8 +7155,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x21d6e5b949392186a6510c57da615f086f779b713a7f3d54c82a07d443e85c5d',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:  {
@@ -7187,8 +7193,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x3ea3a5122344ceb335ff776dbfa5f3e5a06b6edd965414f287f0d57f92304c89',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:  {
@@ -7223,8 +7231,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x1255b3b013addc93fb0301ad2a4150d15e6c1a1c4badbc653fc821b127a37929',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:  {
@@ -7259,8 +7269,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x2c96b715a2e2fe5602d518d75bc4912e6be16f34d9f1f2f420ff6e6de40c9379',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction: {
@@ -7295,8 +7307,10 @@ const transactionByLock = [
     },
     txStatus: {
       blockHash: '0x02bfadbf60172d77af71ad141536d14f5ba191b40b5cffb2b2e7905459e8d500',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
 ]
 
@@ -7381,8 +7395,10 @@ const transactionsByType = [
     },
     txStatus: {
       blockHash: '0x4cc7b42c12e0ed1c87c3ced726e419ba19f755be1739097d6758b6bf60c654ad',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   },
   {
     transaction:       {
@@ -7445,8 +7461,10 @@ const transactionsByType = [
     },
     txStatus: {
       blockHash: '0xd4b10e5af3dac133888f47baeda057f7760fb4f81b2f4dc03a29c228c7dba7a0',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   }
 ];
 
@@ -7512,8 +7530,10 @@ const transactionsByLockAndType = [
     },
     txStatus: {
       blockHash: '0xd4b10e5af3dac133888f47baeda057f7760fb4f81b2f4dc03a29c228c7dba7a0',
-      status: 'pending',
-    }
+      status: 'committed',
+    },
+    cycles: null,
+    timeAddedToPool: null
   }
 ];
 
