@@ -1,7 +1,9 @@
+import { HashType } from "@ckb-lumos/base";
+
 /** Deployed script on chain */
 export interface ScriptConfig {
   CODE_HASH: string;
-  HASH_TYPE: "type" | "data" | "data1";
+  HASH_TYPE: HashType;
   TX_HASH: string;
   INDEX: string;
   DEP_TYPE: "depGroup" | "code";
