@@ -1,27 +1,10 @@
-// string or { branchName: string, label: string }
-const branches = [
-  {
-    branchName: "stable",
-    label: "0.19.x(latest)",
-  },
-  {
-    branchName: "develop",
-    label: "0.20.x(next)",
-  },
+/**
+ *
+ * @type {{href: string, label: string, target: string}[]}
+ */
+module.exports = [
+  { label: "v0.19.0", href: "https://lumos-website-git-019-cryptape.vercel.app/api/", target: "_blank" },
+  { label: "v0.20.0", href: "https://lumos-website-git-publish-0200-magickbase.vercel.app/api/", target: "_blank" },
+  { label: "v0.21.0", href: "https://lumos-website-git-prepare-021-magickbase.vercel.app/api/", target: "_blank" },
+  { label: "canary", href: "https://lumos-website.vercel.app/api/", target: "_blank" },
 ]
-
-const parseURL = (branchName) => branchName.replaceAll("_", "").replaceAll(".", "").replaceAll("/", "-")
-
-module.exports = branches.map((branch) =>
-  typeof branch === "string"
-    ? {
-        href: `https://lumos-website-git-${parseURL(branch)}-magickbase.vercel.app/api/`,
-        label: branch,
-        target: "_blank",
-      }
-    : {
-        href: `https://lumos-website-git-${parseURL(branch.branchName)}-magickbase.vercel.app/api/`,
-        label: branch.label,
-        target: "_blank",
-      }
-)
